refactor: extract further parts for mapTreeViewLevelItemContent
refs #2318
