+++ visualization/app/codeCharta/ui/mapTreeView/mapTreeView.module.ts
@@ -11,7 +11,7 @@ import { mapTreeViewComponent } from "./mapTreeView.component"
 import { mapTreeViewLevelComponent } from "./mapTreeView.level.component"
 
 import { MapTreeViewLevelItemIcon } from "./mapTreeViewLevelItemIcon/mapTreeViewLevelItemIcon.component"
-import { MapTreeViewLevelItemContent } from "./mapTreeViewLevelItemIcon/mapTreeViewLevelItemContent.component"
+import { MapTreeViewLevelItemContent } from "./mapTreeViewLevelItemContent/mapTreeViewLevelItemContent.component"
 
 angular
 	.module("app.codeCharta.ui.mapTreeView", ["app.codeCharta.state", "app.codeCharta.ui.codeMap", "app.codeCharta"])
