--- visualization/app/codeCharta/ui/mapTreeView/mapTreeViewLevelItemIcon/mapTreeViewLevelItemContent.component.html
@@ -1,1 +1,0 @@
-<span></span>
