+++ visualization/app/codeCharta/ui/mapTreeView/pipes/isBuildingHovered.pipe.ts
@@ -1,0 +1,10 @@
+import { Pipe, PipeTransform } from "@angular/core"
+
+import { CodeMapNode } from "../../../codeCharta.model"
+
+@Pipe({ name: "isHoveredBuilding" })
+export class IsHoveredBuildingPipe implements PipeTransform {
+	transform(value: CodeMapNode, hoveredBuildingId: string) {
+		return Boolean(value?.path && hoveredBuildingId === value.path)
+	}
+}
